+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Principles of a change log, excerpted from [Keep a Changelog](http://keepachange
 * Modules:
   - Base module (`Adventure`):
     - `Actor` class, for creating NPC descriptions;
+    - `Building` class, for creating general buildings, such as houses;
     - `Settlement` class, for creating settlements and their contents.
 
 ### Changed